+++ functions/lib/index.js
@@ -1,12 +1,14 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.sendDailyReminders = exports.sendPushNotificationOnWrite = exports.deleteAccount = exports.verifyPurchase = exports.resetUsageLimitsNow = exports.resetWeeklyUsageLimits = exports.getNutritionData = exports.migrateHouseholdClaimsHttp = exports.migrateHouseholdClaims = exports.checkInvitation = exports.leaveHouseholdHttp = exports.leaveHousehold = exports.inviteMemberHttp = exports.inviteMember = void 0;
+exports.sendDailyReminders = exports.sendPushNotificationOnWrite = exports.deleteAccount = exports.handlePlaySubscriptionNotification = exports.verifyPurchase = exports.resetUsageLimitsNow = exports.resetWeeklyUsageLimits = exports.getNutritionData = exports.migrateHouseholdClaimsHttp = exports.migrateHouseholdClaims = exports.checkInvitation = exports.removeHouseholdMember = exports.leaveHouseholdHttp = exports.leaveHousehold = exports.inviteMemberHttp = exports.inviteMember = void 0;
 var inviteMember_1 = require("./inviteMember");
 Object.defineProperty(exports, "inviteMember", { enumerable: true, get: function () { return inviteMember_1.inviteMember; } });
 Object.defineProperty(exports, "inviteMemberHttp", { enumerable: true, get: function () { return inviteMember_1.inviteMemberHttp; } });
 var leaveHousehold_1 = require("./leaveHousehold");
 Object.defineProperty(exports, "leaveHousehold", { enumerable: true, get: function () { return leaveHousehold_1.leaveHousehold; } });
 Object.defineProperty(exports, "leaveHouseholdHttp", { enumerable: true, get: function () { return leaveHousehold_1.leaveHouseholdHttp; } });
+var removeHouseholdMember_1 = require("./removeHouseholdMember");
+Object.defineProperty(exports, "removeHouseholdMember", { enumerable: true, get: function () { return removeHouseholdMember_1.removeHouseholdMember; } });
 var checkInvitation_1 = require("./checkInvitation");
 Object.defineProperty(exports, "checkInvitation", { enumerable: true, get: function () { return checkInvitation_1.checkInvitation; } });
 var migrateHouseholdClaims_1 = require("./migrateHouseholdClaims");
@@ -19,6 +21,8 @@ Object.defineProperty(exports, "resetWeeklyUsageLimits", { enumerable: true, get
 Object.defineProperty(exports, "resetUsageLimitsNow", { enumerable: true, get: function () { return resetUsageLimits_1.resetUsageLimitsNow; } });
 var verifyPurchase_1 = require("./verifyPurchase");
 Object.defineProperty(exports, "verifyPurchase", { enumerable: true, get: function () { return verifyPurchase_1.verifyPurchase; } });
+var subscriptionNotifications_1 = require("./subscriptionNotifications");
+Object.defineProperty(exports, "handlePlaySubscriptionNotification", { enumerable: true, get: function () { return subscriptionNotifications_1.handlePlaySubscriptionNotification; } });
 var deleteAccount_1 = require("./deleteAccount");
 Object.defineProperty(exports, "deleteAccount", { enumerable: true, get: function () { return deleteAccount_1.deleteAccount; } });
 var sendPushNotification_1 = require("./sendPushNotification");